# server/pkg/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"io"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// Target defines model for target.
type Target = string

// Version defines model for version.
type Version = string

// BadRequest defines model for BadRequest.
type BadRequest struct {
	Errors *[]struct {
		Message string `json:"message"`
	} `json:"errors,omitempty"`
	Message string `json:"message"`
}

// Conflict defines model for Conflict.
type Conflict struct {
	Message string `json:"message"`
}

// InternalServerErrror defines model for InternalServerErrror.
type InternalServerErrror struct {
	Message string `json:"message"`
}

// NotFound defines model for NotFound.
type NotFound struct {
	Message string `json:"message"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetFirmwareByTargetAndVersion Get firmware by target and version
	// (GET /firmware/{target}/{version})
	GetFirmwareByTargetAndVersion(ctx echo.Context, target Target, version Version) error
	// PublishNewFirmware Publish new firmware
	// (POST /firmware/{target}/{version})
	PublishNewFirmware(ctx echo.Context, target Target, version Version) error
	// PublishNewFrame Publish new frame
	// (POST /frames)
	PublishNewFrame(ctx echo.Context) error
	// GetFrameByVersion Get frame by version
	// (GET /frames/{version})
	GetFrameByVersion(ctx echo.Context, version Version) error
	// GetStatus Get status
	// (GET /status)
	GetStatus(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetFirmwareByTargetAndVersion converts echo context to params.
func (w *ServerInterfaceWrapper) GetFirmwareByTargetAndVersion(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "target" -------------
	var target Target

	err = runtime.BindStyledParameterWithOptions("simple", "target", ctx.Param("target"), &target, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter target: %s", err))
	}

	// ------------- Path parameter "version" -------------
	var version Version

	err = runtime.BindStyledParameterWithOptions("simple", "version", ctx.Param("version"), &version, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter version: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetFirmwareByTargetAndVersion(ctx, target, version)
	return err
}

// PublishNewFirmware converts echo context to params.
func (w *ServerInterfaceWrapper) PublishNewFirmware(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "target" -------------
	var target Target

	err = runtime.BindStyledParameterWithOptions("simple", "target", ctx.Param("target"), &target, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter target: %s", err))
	}

	// ------------- Path parameter "version" -------------
	var version Version

	err = runtime.BindStyledParameterWithOptions("simple", "version", ctx.Param("version"), &version, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter version: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PublishNewFirmware(ctx, target, version)
	return err
}

// PublishNewFrame converts echo context to params.
func (w *ServerInterfaceWrapper) PublishNewFrame(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PublishNewFrame(ctx)
	return err
}

// GetFrameByVersion converts echo context to params.
func (w *ServerInterfaceWrapper) GetFrameByVersion(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "version" -------------
	var version Version

	err = runtime.BindStyledParameterWithOptions("simple", "version", ctx.Param("version"), &version, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter version: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetFrameByVersion(ctx, version)
	return err
}

// GetStatus converts echo context to params.
func (w *ServerInterfaceWrapper) GetStatus(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetStatus(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(options.BaseURL+"/frames", wrapper.PublishNewFrame, options.OperationMiddlewares["publishNewFrame"]...)
	router.GET(options.BaseURL+"/status", wrapper.GetStatus, options.OperationMiddlewares["getStatus"]...)
	router.GET(options.BaseURL+"/frames/:version", wrapper.GetFrameByVersion, options.OperationMiddlewares["getFrameByVersion"]...)
	router.GET(options.BaseURL+"/firmware/:target/:version", wrapper.GetFirmwareByTargetAndVersion, options.OperationMiddlewares["getFirmwareByTargetAndVersion"]...)
	router.POST(options.BaseURL+"/firmware/:target/:version", wrapper.PublishNewFirmware, options.OperationMiddlewares["publishNewFirmware"]...)

}

type BadRequestJSONResponse struct {
	Errors *[]struct {
		Message string `json:"message"`
	} `json:"errors,omitempty"`
	Message string `json:"message"`
}

type ConflictJSONResponse struct {
	Message string `json:"message"`
}

type InternalServerErrrorJSONResponse struct {
	Message string `json:"message"`
}

type NotFoundJSONResponse struct {
	Message string `json:"message"`
}

type GetFirmwareByTargetAndVersionRequestObject struct {
	Target  Target  `json:"target"`
	Version Version `json:"version"`
}

type GetFirmwareByTargetAndVersionResponseObject interface {
	VisitGetFirmwareByTargetAndVersionResponse(w http.ResponseWriter) error
}

type GetFirmwareByTargetAndVersion200JSONResponse struct {
	Data map[string]interface{} `json:"data"`
}

func (response GetFirmwareByTargetAndVersion200JSONResponse) VisitGetFirmwareByTargetAndVersionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetFirmwareByTargetAndVersion400JSONResponse struct{ BadRequestJSONResponse }

func (response GetFirmwareByTargetAndVersion400JSONResponse) VisitGetFirmwareByTargetAndVersionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetFirmwareByTargetAndVersion404JSONResponse struct{ NotFoundJSONResponse }

func (response GetFirmwareByTargetAndVersion404JSONResponse) VisitGetFirmwareByTargetAndVersionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetFirmwareByTargetAndVersion500JSONResponse struct {
	InternalServerErrrorJSONResponse
}

func (response GetFirmwareByTargetAndVersion500JSONResponse) VisitGetFirmwareByTargetAndVersionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type PublishNewFirmwareRequestObject struct {
	Target  Target  `json:"target"`
	Version Version `json:"version"`
}

type PublishNewFirmwareResponseObject interface {
	VisitPublishNewFirmwareResponse(w http.ResponseWriter) error
}

type PublishNewFirmware201JSONResponse struct {
	Data map[string]interface{} `json:"data"`
}

func (response PublishNewFirmware201JSONResponse) VisitPublishNewFirmwareResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type PublishNewFirmware400JSONResponse struct{ BadRequestJSONResponse }

func (response PublishNewFirmware400JSONResponse) VisitPublishNewFirmwareResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PublishNewFirmware409JSONResponse struct{ ConflictJSONResponse }

func (response PublishNewFirmware409JSONResponse) VisitPublishNewFirmwareResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type PublishNewFirmware500JSONResponse struct {
	InternalServerErrrorJSONResponse
}

func (response PublishNewFirmware500JSONResponse) VisitPublishNewFirmwareResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type PublishNewFrameRequestObject struct {
	Body io.Reader
}

type PublishNewFrameResponseObject interface {
	VisitPublishNewFrameResponse(w http.ResponseWriter) error
}

type PublishNewFrame201JSONResponse struct {
	Data map[string]interface{} `json:"data"`
}

func (response PublishNewFrame201JSONResponse) VisitPublishNewFrameResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type PublishNewFrame400JSONResponse struct{ BadRequestJSONResponse }

func (response PublishNewFrame400JSONResponse) VisitPublishNewFrameResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PublishNewFrame500JSONResponse struct {
	InternalServerErrrorJSONResponse
}

func (response PublishNewFrame500JSONResponse) VisitPublishNewFrameResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetFrameByVersionRequestObject struct {
	Version Version `json:"version"`
}

type GetFrameByVersionResponseObject interface {
	VisitGetFrameByVersionResponse(w http.ResponseWriter) error
}

type GetFrameByVersion200JSONResponse struct {
	Data map[string]interface{} `json:"data"`
}

func (response GetFrameByVersion200JSONResponse) VisitGetFrameByVersionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetFrameByVersion400JSONResponse struct{ BadRequestJSONResponse }

func (response GetFrameByVersion400JSONResponse) VisitGetFrameByVersionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetFrameByVersion404JSONResponse struct{ NotFoundJSONResponse }

func (response GetFrameByVersion404JSONResponse) VisitGetFrameByVersionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetFrameByVersion500JSONResponse struct {
	InternalServerErrrorJSONResponse
}

func (response GetFrameByVersion500JSONResponse) VisitGetFrameByVersionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetStatusRequestObject struct {
}

type GetStatusResponseObject interface {
	VisitGetStatusResponse(w http.ResponseWriter) error
}

type GetStatus200JSONResponse struct {
	Data map[string]interface{} `json:"data"`
}

func (response GetStatus200JSONResponse) VisitGetStatusResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetStatus400JSONResponse struct{ BadRequestJSONResponse }

func (response GetStatus400JSONResponse) VisitGetStatusResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetStatus404JSONResponse struct{ NotFoundJSONResponse }

func (response GetStatus404JSONResponse) VisitGetStatusResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetStatus500JSONResponse struct {
	InternalServerErrrorJSONResponse
}

func (response GetStatus500JSONResponse) VisitGetStatusResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetFirmwareByTargetAndVersion Get firmware by target and version
	// (GET /firmware/{target}/{version})
	GetFirmwareByTargetAndVersion(ctx context.Context, request GetFirmwareByTargetAndVersionRequestObject) (GetFirmwareByTargetAndVersionResponseObject, error)
	// PublishNewFirmware Publish new firmware
	// (POST /firmware/{target}/{version})
	PublishNewFirmware(ctx context.Context, request PublishNewFirmwareRequestObject) (PublishNewFirmwareResponseObject, error)
	// PublishNewFrame Publish new frame
	// (POST /frames)
	PublishNewFrame(ctx context.Context, request PublishNewFrameRequestObject) (PublishNewFrameResponseObject, error)
	// GetFrameByVersion Get frame by version
	// (GET /frames/{version})
	GetFrameByVersion(ctx context.Context, request GetFrameByVersionRequestObject) (GetFrameByVersionResponseObject, error)
	// GetStatus Get status
	// (GET /status)
	GetStatus(ctx context.Context, request GetStatusRequestObject) (GetStatusResponseObject, error)
}

type StrictHandlerFunc func(ctx echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetFirmwareByTargetAndVersion operation middleware
func (sh *strictHandler) GetFirmwareByTargetAndVersion(ctx echo.Context, target Target, version Version) error {
	var request GetFirmwareByTargetAndVersionRequestObject

	request.Target = target
	request.Version = version

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetFirmwareByTargetAndVersion(ctx.Request().Context(), request.(GetFirmwareByTargetAndVersionRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetFirmwareByTargetAndVersion")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetFirmwareByTargetAndVersionResponseObject); ok {
		return validResponse.VisitGetFirmwareByTargetAndVersionResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PublishNewFirmware operation middleware
func (sh *strictHandler) PublishNewFirmware(ctx echo.Context, target Target, version Version) error {
	var request PublishNewFirmwareRequestObject

	request.Target = target
	request.Version = version

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PublishNewFirmware(ctx.Request().Context(), request.(PublishNewFirmwareRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PublishNewFirmware")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PublishNewFirmwareResponseObject); ok {
		return validResponse.VisitPublishNewFirmwareResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PublishNewFrame operation middleware
func (sh *strictHandler) PublishNewFrame(ctx echo.Context) error {
	var request PublishNewFrameRequestObject

	request.Body = ctx.Request().Body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PublishNewFrame(ctx.Request().Context(), request.(PublishNewFrameRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PublishNewFrame")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PublishNewFrameResponseObject); ok {
		return validResponse.VisitPublishNewFrameResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetFrameByVersion operation middleware
func (sh *strictHandler) GetFrameByVersion(ctx echo.Context, version Version) error {
	var request GetFrameByVersionRequestObject

	request.Version = version

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetFrameByVersion(ctx.Request().Context(), request.(GetFrameByVersionRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetFrameByVersion")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetFrameByVersionResponseObject); ok {
		return validResponse.VisitGetFrameByVersionResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetStatus operation middleware
func (sh *strictHandler) GetStatus(ctx echo.Context) error {
	var request GetStatusRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetStatus(ctx.Request().Context(), request.(GetStatusRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetStatus")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetStatusResponseObject); ok {
		return validResponse.VisitGetStatusResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}